square_api_project/backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
requests==2.26.0
python-dotenv==0.19.0
square==17.1.0.20211020
